1. Binary Tree AI 16.9.2021: Implemented and it works.
The question on my mind is if the binary tree is really needed for this? In terms of functionality, the thing would collapse, but on the principal level, this is basically FSM, which uses the binary tree to determine the state of an AI agent.

The code isn't perfect on every level, but this is a good base for what I am gonna do with a game engine.